Consolidate utils
